Support for floating point file sizes and times added

+++ CloudSimEx/modules/cloudsim/src/main/java/org/cloudbus/cloudsim/File.java
@@ -59,7 +59,7 @@ public class File {
 	 *             <li>the file size is zero or negative numbers
 	 *             </ul>
 	 */
-	public File(String fileName, int fileSize) throws ParameterException {
+	public File(String fileName, Double fileSize) throws ParameterException {
 		if (fileName == null || fileName.length() == 0) {
 			throw new ParameterException("File(): Error - invalid file name.");
 		}
@@ -214,7 +214,7 @@ public class File {
 	 * 
 	 * @return the file size (in MBytes)
 	 */
-	public int getSize() {
+	public Double getSize() {
 		return attribute.getFileSize();
 	}
 
@@ -223,7 +223,7 @@ public class File {
 	 * 
 	 * @return the file size (in bytes)
 	 */
-	public int getSizeInByte() {
+	public Double getSizeInByte() {
 		return attribute.getFileSizeInByte();
 	}
 
@@ -233,7 +233,7 @@ public class File {
 	 * @param fileSize the file size (in MBytes)
 	 * @return <tt>true</tt> if successful, <tt>false</tt> otherwise
 	 */
-	public boolean setFileSize(int fileSize) {
+	public boolean setFileSize(Double fileSize) {
 		return attribute.setFileSize(fileSize);
 	}
